requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
bcrypt==4.2.0
click==8.1.7
colorama==0.4.6
DateTime==5.5
Django==5.1
ecdsa==0.19.0
fastapi==0.112.2
greenlet==3.0.3
h11==0.14.0
idna==3.8
passlib==1.7.4
psycopg2==2.9.9
pyasn1==0.6.1
pydantic==2.8.2
pydantic-settings==2.5.2
pydantic_core==2.20.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.2
rsa==4.9
setuptools==74.1.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
sqlparse==0.5.1
starlette==0.38.2
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
zope.interface==7.0.3
